+++ src/main/java/com/pwc/ecasofond/repository/EntryRepository.java
@@ -3,5 +3,6 @@ package com.pwc.ecasofond.repository;
 import com.pwc.ecasofond.model.Entry;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EntryRepository extends CrudRepository<Entry, Integer> {
+public interface EntryRepository extends CrudRepository<Entry, Long> {
+
 }
